Development of Office Management System frontend layouts with a computational Content Management System, resulted System will help in the automation of the office tasks and give an efficient approach. HTML, CSS will be used for frontend while for the purpose of validation both server side and client side JavaScript and PHP will be used. MySQL as database service with the help of PHP language will be used in the backend.
Office Management System which  help to create a profile of the different group of people, and let them perform various actions on available data like collect, store, and update, present and calculate, even if the user is not at office campus. The system will help employees to efficiently compute office related data in order to accomplish the specified task. Apart from this, the system will also record the user activities. Here our approach is to use responsive layout development. This application has two major parts: the first part begins with login & registration which helps to maintain role based accessibility and provide profile management facilities. The second one is Task management & Documents management. The Task management part helps the user to keep & track record of various tasks related to him. One can add the new task, assign tasks to other employees, update task, view task summary. Documents management part helps users to add/upload and search documents efficiently.
HTML, CSS will be used for the front end, while for the purpose of validation both server side and client side JavaScript and PHP will be used. MySQL as database service with the help of PHP language will be used in the backend. Following sections describe the main features of the technology platform.

The scope of Work embraces the development of frontend layouts with a computational Content Management System. Resulted System will help in the automation of the office tasks and give an efficient approach. HTML, CSS will be used for frontend while for the purpose of validation both server side and client side JavaScript and PHP will be used. MySQL as database service with the help of PHP language will be used in the backend.